add node pricing and use multiple strategies
This adds a node pricing strategy, intended to provide baseline metrics
for node costs. The motivation is to provide visibility into how much
nodes cost over time independent from the pods that are running on said
node.

An additional motivation is that we should be able to easily sum node
costs to get a fairly accurate idea of overall cluster cost. We should
be able to notice discrepancies if the total cost of all pod-incurred
costs do not add up to the node cost.

Finally, this node pricing metric can be graphed alongside the
WeightedPricingStrategy incurred cost metrics for individual pods since
we often care not just about absolute numbers, but also proportions.