pubsub cost exporting functionality
Adds the `aggregate` command as well as refactoring how we export data.
We now have the concept of the a general cost exporter, with two
implementations: prometheus and pubsub.

This implementation will be quite voluminous in that amount of data it
produces. It makes sense for the system to do some form of buffering
/ aggregation before exporting to the pubsub queue specifically when
using the pubsub exporter. This kind of machinery is already built in to
the prometheus exporter and is less useful there.