measure and tune kostanza to lag differences
It's possible for kostanza to fall behind the desired polling interval
chosen for cost metrics. A simple solution is to observe it when it
happens and tune the data it generates accordingly while providing some
visibility into the occurrence in order to better tune it.

The `kostanza_lag` metric will show the time difference in milliseconds
between the desired time since last calculation and the actual time
since last calculation. For example, if this falls out of spec by 100ms
and you intended to poll every five seconds you may want to notify the
operator.